issue #8 add NumberValGenerationService

+++ core/src/main/java/me/chanjar/oas/server/validator/core/value/schema/NumberVal.java
@@ -1,12 +1,14 @@
 package me.chanjar.oas.server.validator.core.value.schema;
 
+import java.math.BigDecimal;
+
 /**
  * doc: <a href="https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#dataTypeFormat">Data Types</a>
  *
  * @see io.swagger.v3.oas.models.media.NumberSchema
  */
-public class NumberVal extends SchemaVal<Integer> {
-  public NumberVal(Integer value) {
+public class NumberVal extends SchemaVal<BigDecimal> {
+  public NumberVal(BigDecimal value) {
     super(value);
   }
 }
